Setup a basic Sphinx build
I ran sphinx-quickstart and then made slight tweaks to the output to
suit my preferences.  Dependencies for the build are managed with Conda
because Read The Docs supports that.  I initially started with Pipenv,
which is lighter weight and has a nicer UX, but then realized RTD
integration would be harder.  Besides, plenty of people on the team like
Conda and use it already.