use ncov master for submodule;
instead of converting the index
page to .rst, this just creates
a redundant index page in order
to use the latest master branch
from the ncov repo. Later, it
would be nice to either move
the index page into the docs repo
so that it can be .rst, separate
from the tutorial, or import the
text from the index page into an
.rst index page that also includes
the toctree directive.