Improve organisation of the installation docs
Here we improve the clarity of the installation docs by shifting the
introductory content to the main (index) "Installing Nextstrain"
page. This then naturally leads to the installation pages for various
parts of the nextstrain ecosystem.

The rendered titles of the pages in the sidebar are improved.

The content in the "local installation" page is reduced to remove
duplication between that and the augur/auspice install pages, which
will help avoid divergence going forward.