remove sharing data overview as a subpage;
48004f29391449eb686db0678f3d52dd7b511856 moved this
page up to the top level for the sharing data section
but did not remove its original version, so it shows
up nested underneath itself. I'm assuming this was
not the intention, so I'm removing the page in its
original form and leaving the top level one added
in its place.