add naming_clades from sars-cov-2 tutorial;
this just adds this page to
a toctree instead of having
it be a sort of "floating page"
that bumps you out of your place
in the toc tree when you visit it.
It also creates a warning if not
included in a toctree. The long
term solution will probably be
to generally refactor this
tutorial to fit into docs.nextstrain.org
 / rtd better, since it originally
came from nextstrain.github.io/ncov
(github pages)