Fetch remote documents more quickly
Thread-based concurrency reduces time in I/O wait.  Connection pooling
eliminates the overhead of the initial TCP connection to the same host.
Making the request from within Python instead of shelling out to curl
avoids the overhead of spawning and managing a new process.

A quicker fetch is much nicer when iterating on docs locally and
frequently rebuilding.