Add Zika workflow tutorial content from Augur
Adds a copy of the original Snakemake-based Zika tutorial from the Augur
repository. This is part of an effort to refine the content of the Zika
and TB tutorials to better introduce the concept of bioinformatics
workflows with Snakemake and justify their use. This content provides a
bridge from the first Zika tutorial that focuses on Augur commands on
the shell and the SARS-CoV-2 tutorial that focuses on configuring a
sophisticated Snakemake workflow through YAML files.