Consolidate installation docs into a single page
Adds a new top-level page with the installation instructions that
provide the single clearest path (and an alternate) for most users. The
recommended path uses Conda to bootstrap the user's Python environment
and install all Nextstrain tools. Additional (optional) instructions to
use the Docker image with the Nextstrain CLI follow this initial
installation.