Fork install docs into Conda vs. Docker earlier
Reorganizes the installation docs such that all installations first
install and update Conda. Then, the user is given a choice to install
either the full suite of Nextstrain tools with Conda or to install these
tools via the CLI and Docker. By forking the Conda and Docker options
earlier, the instructions for either approach can be more self-contained
and clearer. By always starting with a Conda installation, we can ensure
that the user's Python environment is properly bootstrapped to a recent
version of Python 3.