Clarify purpose of installing Conda first and update instructions
Clarifies instructions to explain why Conda needs to be installed before
either the "Install with Conda" or "Install with Docker" approaches.
Also, updates the conda commands for updating the environment to make
them simpler and fix a bug in the original update of the Nextstrain
conda environment.

Since the `nextstrain` meta package does not pin any versions and does
not get updated itself very often, running `conda update nextstrain` or
even `conda update --all` does not update Nextstrain dependencies like
Augur the way it should. Instead, we need to explicitly ask to update
the dependencies of the `nextstrain` meta package.