Group install/update instructions into tabs
Uses the sphinx-tabs package to group installation and update
instructions into tabs for Conda or Docker. This commit uses the
"grouped tab" approach such that selecting "Docker" in the installation
section automatically selects "Docker" in the update section, too.

These changes should make clear to the user that Conda and Docker
approaches are mutually exclusive.