Conda install from individual packages instead of nextstrain meta package
The goal of the `nextstrain` meta package in Bioconda [1] was to create
a unified starting point for users to install all of the packages from
our Nextstrain ecosystem with a simpler install command (e.g., `conda
install nextstrain`). We designed this meta package to not pin any
specific versions of dependencies, with the hope that the Conda solvers
would be able to pull the latest versions of all dependencies as needed.

In practice, however, both Conda's and Mamba's solvers default to
installing older versions of Augur even when new versions are available.
This specific issue may be caused by Augur's Bioconda dependency on
Snakemake [2] that the Augur Python package itself does not have [3] and
interactions with Snakemake's other dependencies. Whatever the root
cause of the issue, we most likely cannot use the `nextstrain` meta
package without pinning specific versions of all dependencies and then
updating those manually with separate PRs with each new release of those
dependencies.

To avoid this additional maintenance overhead, this commit replaces
references to the `nextstrain` meta package with the specific list of
packages that are included in that meta package. I've verified that this
approach installs the most recent versions of all packages (as expected)
and also updates to the latest versions using the simpler `conda update
--all` command. We should revisit the meta package maintenance, because
it is likely to be a pain to maintain this manual list of packages in
the future. For now though, these docs work well enough to be merged.

[1]
https://github.com/bioconda/bioconda-recipes/blob/master/recipes/nextstrain/meta.yaml
[2]
https://github.com/bioconda/bioconda-recipes/blob/38d9463b096b5789ed33a7dc0293afef8bb2dbe8/recipes/augur/meta.yaml#L31
[3] https://github.com/nextstrain/augur/blob/15dfc8bc5c5a570066ddd5601c9aa82e0c1d2b7d/setup.py#L49-L57