Include Snakemake as an explicit dependency
Adds Snakemake to the list of packages to install along with Augur,
Auspice, and the CLI in anticipation of removing Snakemake as a
dependency of the Augur Bioconda package. Augur itself does not require
Snakemake (anymore) and this dependency in the Bioconda package is an
oversight on my part from when we unified Augur dependencies between
PyPI and Bioconda.

If we decide to continue using the `nextstrain` meta package, we would
add Snakemake as a dependency of that package.