Attempt to fix the error of existing group and/or user on mac
The error was
```
addgroup: The GID `20' is already in use
```

Turns out nowadays Docker for Mac and Docker for Windows translate UID and GID automagically. And apparently it makes it so that the GID (and perhaps UID?) already exist in the container. This is not the case on Linux.

All Dockerfiles relying on consistent UID and GID across platforms are now broken. And it's the only good way (not mentioning half-baked rootles mode that noone ever bothers to confugure) to get docker containers to not run as root and to not create bunch of root-owned files in volumes. What a wonderful decision!

So maybe if we try to only create the group and the user if not already exist,  this will avoid errors? I cannot test it because I don't have a mac.

See: https://stackoverflow.com/questions/43097341/docker-on-macosx-does-not-translate-file-ownership-correctly-in-volumes