Give the lay of the land on the documentation home page
Full rewrite, starting from some (less extensive) notes I wrote up
earlier this year describing the different pieces of Nextstrain to
someone.

The intent is to help people construct a mental model of the Nextstrain
ecosystem so they'll find it easier to navigate the docs and know what
they're looking for.  Each paragraph tries to build upon the previous
one to link together terminology, with hyperlinks to the topics
mentioned for immediate jumping off points and further reading.