Consistently link to the "stable" version of subproject docs
When available.  The ncov subproject currently only has a "latest"
version.  Stable is a friendlier version to link to as it avoids issues
with documentation being ahead of released code.

If toctree supported Intersphinx links, then we wouldn't have these URLs
scattered around; they'd all be in src/conf.py.