guides/share/index: Replace named links with anonymous links
Fixes a warning about two different `this page <…>`_ links.

In rST, named links (a form of ref) use a single trailing underscore
like `foo <bar>`_ and can be referred to again later by writing foo_.
The names must be unique.  Anonymous links use two trailing underscores
and cannot be referred to again.  An anonymous link is typically more
appropriate when a target path/URL is provided inline.  If the target
path/URL is to be referred to multiple times, using named links which
point to a single definition is better.

I suspect this same mistake is present in other documents of ours, but I
didn't audit for it.