Rough out a learning ramp and improve wayfinding deeper into the docs
I attempted to make a clearer paths deeper into the docs and better
learning ramp starting from the landing page.  That page is now a
junction that provides a brief introduction and signage for where to go
next.  The main ramp is:

  ✦ Landing page
  ↓ What is Nextstrain?
  ↓ Parts of a whole
  ↓ Install
  ↓ Running a build
  ↓ Exploring Zika evolution

The new /learn/about page, now titled "What is Nextstrain?",
describes Nextstrain from a high-level: website + software + platform.
It draws a lot of its language and structure from the previous "About
Nextstrain" page and the nextstrain.org home page.  Intent is that
this is the place for the "Nextstrain manifesto", so to speak.
Previous descriptions of Augur vs. Auspice on the old page were
removed in favor of /learn/parts.

The new /learn/parts page, titled "Parts of a whole" (not great, I
know), has the explanatory text that I previously wrote for the landing
page.  As I was reconciling some overlap of what I wrote with the
previous about page, I realized this "parts" page better fits as the
second or third thing someone reads rather than the first thing.  Some
visitors just interested in understanding the website won't have a need
to read it at all.  I've tried to standardize on "parts" for describing
Nextstrain's many pieces or components, as "parts" seems both more
approachable/less technical and not as overloaded as "components" is in
the software space.

The /tutorials/quickstart content remains the same, but is now titled
"Running a build" and is referred to as a tutorial everywhere instead of
a quickstart.  This makes more sense with its position along the
learning ramp as a short tutorial with the sole purpose of introducing
builds and getting people some quick accomplishments before diving into
more details about builds.