Install docutils<0.17 properly using conda
docutils=0.17 has issues rendering lists [1] so we want to install an
earlier version.

This is only possible if sphinx-tabs is installed with conda, otherwise
pip would resolve the latest sphinx-tabs which requires docutils=0.17
as of version 3.3.0 [2] and re-install docutils=0.17 over any existing
version (third example here [3]).

When sphinx-tabs is installed with conda, conda will resolve the
dependency according to the docutils<0.17 pin which pulls an older
version of sphinx-tabs. This is necessary until sphinx-tabs stops
requiring docutils=0.17.

Additionally, change the conda channel to conda-forge since defaults
does not provide sphinx-tabs.

[1] https://stackoverflow.com/q/67542699
[2] https://github.com/executablebooks/sphinx-tabs/releases/tag/v3.3.0
[3] https://github.com/nextstrain/ncov/issues/891#issuecomment-1067560156