guides/share: Link from summary table to the summary sections on the same page
Otherwise the table jumps you to other pages or even other sites and
you're likely to miss the brief summaries on this page.  Each summary
section now ends with the link out to other pages/sites for more
information.

A section was missing for "Public URLs", so I added one.