guides/contribute: Link out to project-specific contributing guides instead of transcluding them
The best context for these documents is the repository they live in, not
the documentation website.  The documents are written (reasonably so!)
to be viewed/used alongside the code itself.  Linking to the docs on
GitHub helps facilitate this context and code browsing.

This also reduces the number of documents we're transcluding, which I
see as a bonus since long-term I think we want to stop this practice (or
alternatively, double down on it and start doing it much more
systematically (e.g. with cross-project link resolution/rewriting)).

Initially motivated by CI failing because a relative link to Augur's
changelog was added to Augur's development guide.  Such an addition is
completely reasonable; the naive transclusion was not.