Add commands to create intermediate conda environment with osx-64 emulation
This is for users on Apple Silicon. If emulation is not used, Conda will fail to find packages like mafft and iqtree since they are not available via osx-arm64 or noarch. This is outside of our control, so the best recommendation is to use emulation which has worked well for me personally.