Add note to look for other specific installation steps
Users may run through this generic Nextstrain installation and skip over
workflow-specific setup guides, which are necessary for installing
workflow-specific software requirements (e.g. pangolin for SARS-CoV-2)
that are not included in the generic installation steps.

This change adds a note and example link to SARS-CoV-2 setup steps given
the currently high usage of the SARS-CoV-2 workflow.

Note that this only impacts the Native environment.