Install more Python deps with Pip instead of Conda
This is already done in https://github.com/nextstrain/ncov/commit/f882543dbcd73b0251d39ddcfa9516380eecedbb
and the same issues with the docs build are showing here.

From the commit linked above:

Everything but make is pip-installable and the Conda environment is
currently having issues on both RTD (where builds are stopped due to
memory usage) and on GitHub Actions (where builds fail in pip during dep
resolution during the Conda env install).¹  Installing more things with
Pip may help it out.

¹ https://bedfordlab.slack.com/archives/C01LCTT7JNN/p1657059168195909

The only difference here is that graphviz on pip is not the same as
graphviz on conda, so the conda dependency should be retained.

Co-authored-by: Thomas Sibley <tsibley@fredhutch.org>