install: Use --strict-channel-priority for Conda channels
This is a documented requirement by both conda-forge¹ and bioconda² so
that packages from the "defaults" channel get selected only if they're
not in either of the others.  It is also required to solve a dependency
resolution issue affecting Nextstrain CLI installed from Bioconda that
arises because of the otherwise "flexible" channel priority.³

While conda-forge and bioconda suggest setting this in .condarc with
`conda config`, I've used the command-line option instead to match our
convention of specifying channels on the command-line too.

Note that specifying `-c defaults` explicitly isn't necessary most of
the time because it should already be last priority in the channel list,
but it's possible folks have used .condarc to remove it entirely due to
outdated advice.

¹ https://conda-forge.org/docs/user/tipsandtricks.html#using-multiple-channels
² https://bioconda.github.io/#usage
³ https://github.com/nextstrain/cli/issues/184