install: Distinguish between Bash and PowerShell
Currently no syntax differences between the code blocks, but there will
be, for example, once we break a single command across lines across
lines as they have different line continuation characters (\ vs. `).
Distinguishing them also helps us when authoring to remember the context
in which the commands are to be run, e.g. to avoid absentmindedly using
bash-isms in PowerShell.