install: Switch to persistently configuring Conda channels and settings
…instead of specifying command-line options to every invocation.  This
simplifies the commands and makes later installs/updates more foolproof
since the required command-line options can't be forgotten.

The Conda config is set for the "nextstrain" environment only so as not
to affect other envs, which means changing the order of operations from

  (create + install) → activate

to

  create → activate → configure → install.

The snippets for the minimal Conda env with just nextstrain-cli are now
inlined since the command is now a short four words long.