Use standalone installation of Nextstrain CLI for Docker runtimes
The goal is simplify this installation path quite a bit by avoiding the
Python bootstrapping issue.

Reorganizes the tabs to be runtime first then OS (instead of OS then
runtime) as I think this makes more sense now that the initial steps of
installing and setting up a base Conda env aren't shared.  It also means
we don't have multiple levels of ordered lists (steps), which is a nice
simplification.

Moves verbose explanatory information out into the FAQ and adds more
explanation about how to make various choices.

Moves the hint for WSL on Windows installs next to the hint for native
runtime installs at the bottom.  This avoids duplicating it in both of
the WSL on Windows tabs and it taking up a lot of visual space in those
tabs.

Related-to: <https://github.com/nextstrain/nextstrain.org/pull/582>
Related-to: <https://github.com/nextstrain/cli/pull/217>