install: Standardize steps for WSL + Docker
These are largely the same whether installing the CLI on Windows or WSL,
and it's useful to see that similarity when flipping between tabs (i.e.
as someone deciding between methods might do).

This also puts the potential machine restart early and thus avoids
having to remember to reactivate the Conda env afterwards.