groups/migrate-from-s3: Simplify upload example
Make the old vs. new comparison more direct by uploading to
nextstrain.org in a way that uses the local filename.

The "/ncov/${YOUR_BUILD_NAME}" bit was supposed to show the parallel to
"ncov_${YOUR_BUILD_NAME}" in the filename.  In retrospect, however, the
psuedo-placeholder syntax will probably only serve to confuse since
"${YOUR_BUILD_NAME}" must be slash-delimited in the former but
underscore-delimited in the latter.