faq: Stop stylizing runtime names as lowercase literals
We don't use this style in other documentation (here and in the CLI's
docs).  I also think it's confusing with the common use of literals for
commands, e.g. our usage of ``conda`` to refer to the actual command and
Conda to the ecosystem.

I've used double quotes instead in places where where some sort of
demarcation from surrounding text felt warranted to avoid ambiguity
(e.g. especially in initial occurrences).  This has some precedent in
other documentation around runtime names as well.