augur-to-auspice: update default colorings type
In responding to a discussion post¹, I found that augur export will try
to guess the appropriate color type if the type is not provided².
This commit updates the augur-to-auspice docs to outline this guessing process.

Also added a link to the guide on how to add custom trait colors for
users to learn more about the colors TSV file.

¹ https://discussion.nextstrain.org/t/zero-value-for-coloring-in-nextstrain/1266/1
² https://github.com/nextstrain/augur/blob/8b3c33dd8994ffeccd6e82ba9c06d2cf0499c3e4/augur/export_v2.py#L245-L251

Co-authored-by: Victor Lin <13424970+victorlin@users.noreply.github.com>