toc: Add "Components" section for our subproject docs
…moving them out of the general "Reference" section.  I think this makes
the table of contents clearer by not mixing links out to subprojects,
each with their own large set of pages, with the other general reference
pages.

I chose the term "components" for the TOC section because it seemed most
descriptive in the isolation of the sidebar.  We (I) intentionally use
the simpler word "parts" in other places, most notably the "Parts of a
whole" page, where there is space to elaborate.  In isolation, though,
"parts" doesn't seem as informative to me.  We also use the word "tools"
in many places in the docs (and colloquially).  That could work here
instead of "components", but to me "components" carries a connotation of
parts of a larger thing, and "tools" doesn't quite have that.