augur-to-auspice: Update section levels and contents depth
Updating the existing section levels for the Auspice section of the
document in preparation for adding a new section for the diversity panel.

This commit also updates the depth of the table of contents for this
page so that it's easier to navigate to relevant sections as this
document continues to grow.