augur-to-auspice: Update "How branch mutations are exported"
Renames section to indicate the same data are used for branch mutations
and the diversity panel. Adds more details on how the nucleotide
mutations, amino acid mutations, and genome annotations are exported.

Note that I excluded details on how `augur ancestral` and `augur translate`
work because I think these details will be more appropriate to add to
the Augur docs.