Use sections for each recommendation
This is consistent with the existing documentation style guide and a
similar Snakemake style guide¹ written by @tsibley.

This also allows for the addition of a table of contents, which is done
here.

I looked into making the section links future-proof to wording changes,
however could not find a good solution. The closest is using :ref: roles
which adds separately defined links to to the section titles, but those
links aren't used by the TOC or header links themselves.

¹ https://github.com/blab/styleguide/blob/6aa5d7aa42acfa97e57a5ee05a4175f158502cac/nextstrain-builds.md