install: Update recommendation for Windows options
Refer to the options by "host interface" instead of "operating system"
since, at least to the user, both Windows options are on the Windows OS.
The rename also makes the WSL recommendation more favorable since it
seems like less of a secondary alternative to PowerShell.

Also, update the FAQ text for the new Conda runtime (the previous text
predated the Conda runtime and assumed that ambient was the only
non-Docker runtime).