guides/versions: Use "snapshots" (instead of "versions") and "analyses"
The document already used the term "snapshot" a bit, and I think it's
less ambiguous/overloaded than "version" so use it consistently.
"Analyses" is already in use in other document and section titles, so
use that term in those contexts.

I didn't update the source document file names here, but maybe I should…