Remove snippets/uncurated-ncbi-dataset.rst
The previous commit updates the creating-an-ingest-workflow tutorial
to use a custom Snakemake target to generate the uncurated NCBI Dataset
metadata so we no longer need the separate snippet.

I am still providing the extra commands to generate the uncurated
data in the running-an-ingest-workflow tutorial because I wanted users
to be able to see the uncurated data even if the existing pathogen
workflow does _not_ include the custom Snakemake target.