Update ambiant install docs to specify Python version
I also modified the overall intent of the instructions slightly, so
that they direct the user to set up a new Conda environment. The old
instructions were oriented more towards adding Nextstrain to an
existing Conda enviroment, which (IMO) seems like a less common
scenario, and also one more likely to be undertaken by a more
experienced user who will be able to adapt these instructions on the
fly.