Update WSL home dir snippet
The previous recommendation was to expose the Windows home directory
under the WSL directory. I thought this would be a practical approach to
the difference in home directories, but I'm not a WSL user so I can't
verify that idea. Installing `wslu` is also a bit more complicated
nowadays.¹

Instead of a hard recommendation, simply describe the difference in home
directories and let the user decide what to do.

¹ <https://wslutiliti.es/wslu/install.html#ubuntu>