reference: Document how to configure/modify CA certificate trust stores
This information is a distillation of guidance and assistance I've
provided to CDC and other downstream users over the years.  I extracted
it from my various notes and recollections and verified some specific
behaviour anew.  An official reference page will be useful for pointing
users to in the future, and serves as a collection point for future
recommendations/information uncovered in troubleshooting sessions.

Related-to: <https://github.com/nextstrain/ncov/issues/774>
Related-to: <https://github.com/nextstrain/ncov/issues/1055>
Related-to: <https://github.com/nextstrain/nextclade/issues/726>