Remove paragraph on pipefail with tee
This was meaningful when the recommendation was

    … | tee {log:q}

It no longer makes sense with changes from "Use exec for more robust
logging" (ed66042).

There's still an argument to be made that pipefail should be used with
pipes in general, but that is more about shell scripting and does not
need to be explicitly mentioned given that Snakemake uses bash's strict
mode by default.