Write examples with both augur filter & augur subsample
augur subsample was created with the intent to simplify the process of
subsampling when a single call to augur filter does not suffice, however
it can also be used as a replacement for augur filter entirely.

By introducing augur subsample to users early on, it should be easier
for them to transition from simple to complex subsampling
configurations.